explicit list

+++ mm.h
@@ -2,9 +2,9 @@
 #include <stdint.h>
 
 extern int mm_init(void);
-extern void *mm_malloc(size_t size);
+extern void *mm_malloc(uint32_t size);
 extern void mm_free(void *ptr);
-extern void *mm_realloc(void *ptr, size_t size);
+extern void *mm_realloc(void *ptr, uint32_t size);
 
 /* 
  * Students work in teams of one or two.  Teams enter their team name,
